create store and reducer
