+++ src/store/initialState.js
@@ -1,0 +1,3 @@
+const initialState = {result: 0, showError: false, inputValue: ''}
+
+export default  initialState
